+++ test/analysis/test-sizing-lp.jl
@@ -8,7 +8,7 @@
 end
 
 @testmodule LPHome begin
-    using BatteryBusinessCase
+    using HEMSSimulator
     using Dates: DateTime
 
     site = Site(52.1, 5.18)