+++ test/assets/test-heatpump.jl
@@ -1,5 +1,5 @@
 @testmodule WinterHome begin
-    using BatteryBusinessCase
+    using HEMSSimulator
     using Dates: DateTime
 
     site = Site(52.1, 5.18)