+++ test/test-quality.jl
@@ -1,5 +1,5 @@
 @testitem "Aqua" tags = [:unit, :fast] begin
     using Aqua: Aqua
-    Aqua.test_all(BatteryBusinessCase; ambiguities = false)
-    Aqua.test_ambiguities(BatteryBusinessCase)
+    Aqua.test_all(HEMSSimulator; ambiguities = false)
+    Aqua.test_ambiguities(HEMSSimulator)
 end